requirements.txt
colorama
loguru
pydantic
PyYAML
tabulate
typing-extensions